# ticketing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the TicketPulse API for ticketing microservice. Concerts, sessions, sections and tickets are all here.",
        "title": "TicketPulse Ticketing API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/tickets": {
            "get": {
                "description": "Retrieves a list of all tickets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all Tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Ticket"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/tickets/{id}": {
            "get": {
                "description": "Retrieves a ticket by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Ticket by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a ticket by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update Ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Seat Number",
                        "name": "seatNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User ID",
                        "name": "userId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ent.Concert": {
            "type": "object",
            "properties": {
                "artist": {
                    "description": "Artist holds the value of the \"artist\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ConcertQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ConcertEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "imageUrl": {
                    "description": "ImageUrl holds the value of the \"imageUrl\" field.",
                    "type": "string"
                },
                "title": {
                    "description": "Title holds the value of the \"title\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ConcertEdges": {
            "type": "object",
            "properties": {
                "hasConcertSessions": {
                    "description": "HasConcertSessions holds the value of the hasConcertSessions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ConcertSession"
                    }
                }
            }
        },
        "ent.ConcertSession": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ConcertSessionQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ConcertSessionEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "sessionDateTime": {
                    "description": "SessionDateTime holds the value of the \"sessionDateTime\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.ConcertSessionEdges": {
            "type": "object",
            "properties": {
                "hasSections": {
                    "description": "HasSections holds the value of the hasSections edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Section"
                    }
                },
                "ofConcert": {
                    "description": "OfConcert holds the value of the ofConcert edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Concert"
                    }
                }
            }
        },
        "ent.Section": {
            "type": "object",
            "properties": {
                "bought": {
                    "description": "Bought holds the value of the \"bought\" field.",
                    "type": "integer"
                },
                "capacity": {
                    "description": "Capacity holds the value of the \"capacity\" field.",
                    "type": "integer"
                },
                "category": {
                    "description": "Category holds the value of the \"category\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/section.Category"
                        }
                    ]
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SectionQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.SectionEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "integer"
                },
                "reserved": {
                    "description": "Reserved holds the value of the \"reserved\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.SectionEdges": {
            "type": "object",
            "properties": {
                "atConcertSession": {
                    "description": "AtConcertSession holds the value of the atConcertSession edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ConcertSession"
                        }
                    ]
                },
                "hasTickets": {
                    "description": "HasTickets holds the value of the hasTickets edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Ticket"
                    }
                }
            }
        },
        "ent.Ticket": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TicketQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.TicketEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "seatNumber": {
                    "description": "SeatNumber holds the value of the \"seatNumber\" field.",
                    "type": "integer"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ticket.Status"
                        }
                    ]
                },
                "userId": {
                    "description": "UserId holds the value of the \"userId\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TicketEdges": {
            "type": "object",
            "properties": {
                "withinSection": {
                    "description": "WithinSection holds the value of the withinSection edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Section"
                        }
                    ]
                }
            }
        },
        "section.Category": {
            "type": "string",
            "enum": [
                "CAT1",
                "CAT2",
                "CAT3",
                "CAT4",
                "CAT5"
            ],
            "x-enum-varnames": [
                "CategoryCAT1",
                "CategoryCAT2",
                "CategoryCAT3",
                "CategoryCAT4",
                "CategoryCAT5"
            ]
        },
        "ticket.Status": {
            "type": "string",
            "enum": [
                "AVAILABLE",
                "BOUGHT",
                "RESERVED"
            ],
            "x-enum-varnames": [
                "StatusAVAILABLE",
                "StatusBOUGHT",
                "StatusRESERVED"
            ]
        },
...
        "ticketing_internal_common_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "ticketing_internal_common_types.Ticket": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "seatNumber": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}